Moved getItemTemplate to templates
