+++ packages/group/test/group.test.ts
@@ -23,6 +23,7 @@ import * as utils from "../../common/test/mocks/utils";
 import * as staticDashboardMocks from "../../common/test/mocks/staticDashboardMocks";
 import * as fetchMock from "fetch-mock";
 import * as mockItems from "../../common/test/mocks/agolItems";
+import * as templates from "../../common/test/mocks/templates";
 import * as common from "@esri/solution-common";
 
 // Set up a UserSession to use in all these tests
@@ -58,7 +59,7 @@ afterEach(() => {
 describe("Module `group`: manages the creation and deployment of groups", () => {
   describe("convertItemToTemplate", () => {
     it("should handle error on getGroup", done => {
-      const itemTemplate: common.IItemTemplate = mockItems.getItemTemplate();
+      const itemTemplate: common.IItemTemplate = templates.getItemTemplate();
       itemTemplate.itemId = "abc0cab401af4828a25cc6eaeb59fb69";
       itemTemplate.item = mockItems.getAGOLItem("Group", null);
       itemTemplate.item.tags = [];
@@ -111,7 +112,7 @@ describe("Module `group`: manages the creation and deployment of groups", () => 
     });
 
     it("should handle error on portal getGroup", done => {
-      const itemTemplate: common.IItemTemplate = mockItems.getItemTemplate();
+      const itemTemplate: common.IItemTemplate = templates.getItemTemplate();
       itemTemplate.itemId = "abc0cab401af4828a25cc6eaeb59fb69";
       itemTemplate.item = mockItems.getAGOLItem("Group", null);
       itemTemplate.item.tags = [];
@@ -234,7 +235,7 @@ describe("Module `group`: manages the creation and deployment of groups", () => 
     });
 
     it("should handle a group", done => {
-      const itemTemplate: common.IItemTemplate = mockItems.getItemTemplate();
+      const itemTemplate: common.IItemTemplate = templates.getItemTemplate();
       itemTemplate.itemId = "abc0cab401af4828a25cc6eaeb59fb69";
       itemTemplate.item = mockItems.getAGOLItem("Group", null);
 
@@ -429,7 +430,7 @@ describe("Module `group`: manages the creation and deployment of groups", () => 
       };
       const templateDictionary: any = { user };
 
-      const itemTemplate: common.IItemTemplate = mockItems.getItemTemplate();
+      const itemTemplate: common.IItemTemplate = templates.getItemTemplate();
       itemTemplate.itemId = itemId;
       itemTemplate.type = "Group";
       itemTemplate.item.title = "Dam Inspection Assignments";
@@ -469,7 +470,7 @@ describe("Module `group`: manages the creation and deployment of groups", () => 
       };
       const templateDictionary: any = { user };
 
-      const itemTemplate: common.IItemTemplate = mockItems.getItemTemplate();
+      const itemTemplate: common.IItemTemplate = templates.getItemTemplate();
       itemTemplate.itemId = itemId;
       itemTemplate.type = "Group";
       itemTemplate.item.title = "Dam Inspection Assignments";
@@ -502,7 +503,7 @@ describe("Module `group`: manages the creation and deployment of groups", () => 
       };
       const templateDictionary: any = { user };
 
-      const itemTemplate: common.IItemTemplate = mockItems.getItemTemplate();
+      const itemTemplate: common.IItemTemplate = templates.getItemTemplate();
       itemTemplate.itemId = itemId;
       itemTemplate.type = "Group";
       itemTemplate.item.title = "Dam Inspection Assignments";
@@ -536,7 +537,7 @@ describe("Module `group`: manages the creation and deployment of groups", () => 
       };
       const templateDictionary: any = { user };
 
-      const itemTemplate: common.IItemTemplate = mockItems.getItemTemplate();
+      const itemTemplate: common.IItemTemplate = templates.getItemTemplate();
       itemTemplate.itemId = itemId;
       itemTemplate.type = "Group";
       itemTemplate.item.title = "Dam Inspection Assignments";
@@ -624,7 +625,7 @@ describe("Module `group`: manages the creation and deployment of groups", () => 
     //     const templateDictionary: any = {};
     //     const newItemID: string = "abc8cab401af4828a25cc6eaeb59fb69";
 
-    //     const itemTemplate: IItemTemplate = mockItems.getItemTemplate();
+    //     const itemTemplate: IItemTemplate = templates.getItemTemplate();
     //     itemTemplate.itemId = itemId;
     //     itemTemplate.type = "Group";
     //     itemTemplate.item.title = "Dam Inspection Assignments";
@@ -665,7 +666,7 @@ describe("Module `group`: manages the creation and deployment of groups", () => 
 
   describe("updateGroup", () => {
     it("should handle error", done => {
-      const itemTemplate: common.IItemTemplate = mockItems.getItemTemplate();
+      const itemTemplate: common.IItemTemplate = templates.getItemTemplate();
       itemTemplate.itemId = "abc1cab401af4828a25cc6eaeb59fb69";
       itemTemplate.dependencies = ["abc0cab401af4828a25cc6eaeb59fb69"];
 
@@ -737,7 +738,7 @@ describe("Module `group`: manages the creation and deployment of groups", () => 
     });
 
     it("should share dependencies with group", done => {
-      const itemTemplate: common.IItemTemplate = mockItems.getItemTemplate();
+      const itemTemplate: common.IItemTemplate = templates.getItemTemplate();
       itemTemplate.itemId = "abc1cab401af4828a25cc6eaeb59fb69";
       itemTemplate.dependencies = ["abc0cab401af4828a25cc6eaeb59fb69"];
